+++ src/common/i18n/en.js
@@ -43,9 +43,9 @@ AccountsManager.i18n = {
                 not_allowed: '<p>You are unfortunately not allowed to list the accounts.</p>'
                     +'<p>You can contact your application administrator to ask for the needed habilitation.</p>',
                 role_global: 'Global',
-                role_global_tooltip: 'These are the direct roles assigned to the user at the global level',
+                role_global_tooltip: 'Direct roles assigned to the \'%s\' user at the global level',
                 role_scoped: 'Scoped',
-                role_scoped_tooltip: 'These are the direct roles assigned to the user at the \'%s\' scope level',
+                role_scoped_tooltip: 'Direct roles assigned to the \'%s\' user at the \'%s\' scope level',
                 roles_th: 'Roles',
                 user_notes_th: 'User Notes',
                 username_th: 'Username'